FEAT: Add control panel in drive
