+++ react-ui/src/components/drive/ui/RoverInfo.tsx
@@ -1,14 +1,11 @@
-import Speedometer from "./ui/Speedometer";
-import Knob from './ui/Knob';
-import GPS from './ui/GPS';
+import WheelsContainer from "./WheelsContainer";
+import Speedometer from "./Speedometer";
+import GPS from './GPS';
 import "./styles/RoverInfo.css";
-import WheelsContainer from "./ui/wheel_angle_minimap/components/WheelsContainer";
 
 const RoverInfo: React.FC = () => {
   return (
-    <div className="drive-info-container">
-      <h1 className="drive-info-title">Information</h1>
-    
+    <div className="drive-info-container"> 
       <div className="speedometer-container">
         <h2 className="speed-title">Rover Speed</h2>
         {/* Main speedometer */}
@@ -37,14 +34,8 @@ const RoverInfo: React.FC = () => {
         </div>
       </div>
 
-      <div className="knob-and-wheel-angles-grid">
-        <div className="knob-container">
-          <h2 className="knob-title">Maximum speed</h2>
-          <Knob />
-        </div>
-        <div className="wheels-minimap-container">
-          <WheelsContainer />
-        </div>
+      <div className="wheels-minimap-container">
+        <WheelsContainer />
       </div>
 
       <div className="gps-container">
