+++ react-ui/src/components/drive/ui/styles/Headlight.css
@@ -1,0 +1,41 @@
+.brightness-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.pizza-segment-container {
+  position: relative;
+  width: 200px;
+  height: 200px;
+  display: flex;
+  align-items: flex-end;
+  justify-content: center;
+  overflow: hidden;
+}
+
+.pizza-segment {
+  width: 200px;
+  height: 200px;
+  border-radius: 50%;
+  clip-path: polygon(50% 100%, 20% 0%, 80% 0%);
+  transition: background 0.3s ease;
+}
+
+input[type="range"] {
+  margin-top: 10px;
+  width: 80%;
+  accent-color: #efff5c;
+  border-radius: 5px;
+  transition: background 0.3s ease;
+  outline: none;
+}
+
+input[type="range"]::-range-track {
+  height: 10px;
+  background: #efff5c;
+  border-radius: 5px;
+}
