+++ react-ui/src/components/styles/Navbar.css
@@ -2,7 +2,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  padding: 30px 2.5%;
+  padding: 15px 2%;
   background-color: #24252A;
   position: relative;
 }
